+++ application/views/equipment.php
@@ -1,8 +1,19 @@
+<!-- author: Michaela -->
+
 <div class="row">
-	<div class="span2">
-                <img src="/pix/{image}.jpg" title="{code}"/><br/>
-                code {code}
-                name {name}
-                attributes {attributes}
+    {authors}
+	<div class="span4">
+                
+                <ul>
+                <img src="/assets/img/{EquipmentID}.png" title="{EquipmentID}"/><br/>
+                <li>Code: {EquipmentID}</li>
+                <li>Name: {Name}</li>
+                <li>Category: {Category}</li>
+                <li>Mobility: {Mobility}</li>
+                <li>Range: {Range}</li>
+                <li>Power: {Power}</li>
+                <li>Protection: {Protection}</li>
+                </ul>
 	</div>
+    {/authors}
 </div>